medical-ai-assistant/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
starlette==0.38.6

celery==5.4.0
redis==5.2.0
kombu==5.6.1
billiard==4.2.3
vine==5.1.0

streamlit==1.40.0

pandas==2.2.3
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
requests==2.32.3
orjson==3.11.4
python-json-logger==3.2.1
coloredlogs==15.0.1

langchain==0.3.7
langchain-core==0.3.63
langchain-community==0.3.7
langchain-chroma==0.1.4
langchain-openai==0.2.8
langchain-text-splitters==0.3.8
chromadb==0.5.20
chroma-hnswlib==0.7.6
sentence-transformers==3.0.1
huggingface-hub==0.36.0
tiktoken==0.12.0

ollama==0.3.3
openai==1.54.0
google-generativeai==0.8.3
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-api-core==2.28.1
googleapis-common-protos==1.72.0

opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-util-http==0.59b0

torch==2.9.1
transformers==4.57.3
scikit-learn==1.7.2
scipy==1.16.3
joblib==1.5.2

watchfiles==1.1.1
watchdog==5.0.3
rich==13.9.4
tenacity==9.0.0

